+++ packages/node/src/stellar/block.stellar.ts
@@ -14,7 +14,7 @@ import {
   StellarTransaction,
   StellarTransactionFilter,
 } from '@subql/types-stellar';
-import { Address, scValToNative, xdr } from 'soroban-client';
+import { Address, scValToNative, xdr } from 'stellar-sdk';
 import { stringNormalizedEq } from '../utils/string';
 
 export class StellarBlockWrapped implements StellarBlockWrapper {
@@ -130,51 +130,11 @@ export class StellarBlockWrapped implements StellarBlockWrapper {
         if (!event.topic[i]) {
           return false;
         }
-        if (topic !== event.topic[i].str()) {
+        if (topic !== scValToNative(event.topic[i])) {
           return false;
         }
       }
     }
-
     return true;
-  }
-
-  static decodeScVals(scVal: xdr.ScVal): any {
-    switch (scVal.switch()) {
-      case xdr.ScValType.scvBool():
-        return scVal.b();
-      case xdr.ScValType.scvSymbol():
-        return scVal.sym().toString();
-      case xdr.ScValType.scvU64():
-        return scVal.u64().low;
-      case xdr.ScValType.scvAddress(): {
-        try {
-          return Address.account(
-            scVal.address().accountId().value(),
-          ).toString();
-        } catch (error) {
-          return Address.contract(scVal.address().contractId()).toString();
-        }
-      }
-      case xdr.ScValType.scvString(): {
-        return Buffer.from(scVal.str().toString(), 'base64').toString();
-      }
-      case xdr.ScValType.scvBytes():
-        return scVal.bytes();
-      case xdr.ScValType.scvI128(): {
-        const low = scVal.i128().lo();
-        const high = scVal.i128().hi();
-        return BigInt(low.low) | (BigInt(low.high) << BigInt(32));
-      }
-      case xdr.ScValType.scvMap():
-        return Object.fromEntries(
-          scVal.map()!.map((entry) => {
-            const key = entry.key().sym();
-            return [key, StellarBlockWrapped.decodeScVals(entry.val())];
-          }),
-        );
-      default:
-        return scValToNative(scVal);
-    }
   }
 }