+++ packages/types/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Use types from stellar sdk (#46)
+
 ## [3.2.1] - 2024-02-07
 ### Changed
 - Update `@subql/types-core`